Split library and local_library
