+++ parser_library/src/workspaces/CMakeLists.txt
@@ -17,8 +17,9 @@ target_sources(parser_library PRIVATE
 	file_manager.h
 	file_manager_impl.cpp
 	file_manager_impl.h
-	library.cpp
 	library.h
+	local_library.cpp
+	local_library.h
 	parse_lib_provider.cpp
 	parse_lib_provider.h
 	processor.h
